+++ figpack-manage/src/pages/ManageFigurePage/ManageFigurePage.tsx
@@ -10,6 +10,7 @@ import React, { useCallback, useMemo, useState } from "react";
 import { useAuth } from "../../hooks/useAuth";
 import { useBacklinks } from "../../hooks/useBacklinks";
 import DeleteDialog from "./DeleteDialog";
+import DocumentReferences from "./DocumentReferences";
 import DownloadInstructions from "./DownloadInstructions";
 import FigureDetails from "./FigureDetails";
 import FigureHeader from "./FigureHeader";
@@ -154,6 +155,9 @@ const ManageFigurePage: React.FC = () => {
             backlinks={backlinks?.filter((b) => b.url === figureUrl)}
           />
         )}
+
+        {/* Document References */}
+        <DocumentReferences figureUrl={figureUrl} formatDate={formatDate} />
 
         {/* Figure Preview */}
         <FigurePreview figureUrl={figureUrl} />